enh: implemented DispersionModel super class
