+++ rocketpy/monte_carlo/__init__.py
@@ -1,9 +1,5 @@
-from .mc_aero_surfaces import (
-    McEllipticalFins,
-    McNoseCone,
-    McTail,
-    McTrapezoidalFins,
-)
+from .DispersionModel import DispersionModel
+from .mc_aero_surfaces import McEllipticalFins, McNoseCone, McTail, McTrapezoidalFins
 from .mc_environment import McEnvironment
 from .mc_flight import McFlight
 from .mc_parachute import McParachute
