+++ rocketpy/monte_carlo/mc_flight.py
@@ -1,11 +1,12 @@
-from typing import Any, List, Tuple, Union
+from typing import Any, Tuple, Union
 
-from pydantic import BaseModel, StrictFloat, StrictInt, root_validator
+from pydantic import StrictFloat, StrictInt
 
 from ..Flight import Flight
+from .DispersionModel import DispersionModel
 
 
-class McFlight(BaseModel):
+class McFlight(DispersionModel):
     """TODO: Add description
 
     Parameters
@@ -46,42 +47,3 @@ class McFlight(BaseModel):
         ],
     ] = None
     terminateOnApogee: bool = False
-
-    class Config:
-        arbitrary_types_allowed = True
-
-    @root_validator(skip_on_failure=True)
-    def val_basic(cls, values):
-        """Validates inputs that can be either tuples or lists.
-        Tuples can have either 2 or 3 items. First two must be either float or int,
-        representing the nominal value and standard deviation. Third item must be
-        a string containing the name of a numpy.random distribution function"""
-
-        validate_fields = ["inclination", "heading"]
-        for field in validate_fields:
-            v = values[field]
-            # checks if tuple
-            if isinstance(v, tuple):
-                # checks if first two items are valid
-                assert isinstance(v[0], (int, float)) and isinstance(
-                    v[1], (int, float)
-                ), f"\nField '{field}': \n\tFirst two items of tuple must be either an int or float \n\tFirst item refers to nominal value, and the second to the standard deviation"
-                # extra check for third item if passed
-                if len(v) == 3:
-                    assert isinstance(
-                        v[2], str
-                    ), f"\nField '{field}': \n\tThird item of tuple must be either a string \n\tThe third item must be a string containing the name of a numpy.random distribution function"
-                # all good, sets inputs
-                values[field] = v
-            elif isinstance(v, list):
-                # guarantee all values are valid (ints or floats)
-                assert all(
-                    isinstance(item, (int, float)) for item in v
-                ), f"\nField '{field}': \n\tItems in list must be either ints or floats"
-                # all good, sets inputs
-                values[field] = v
-            else:
-                raise ValueError(
-                    f"\nField '{field}': \n\tMust be either a tuple or list"
-                )
-        return values
